Steam ROM Manager: Disable killing Steam * Steam ROM Manager can now handle killing and restarting Steam natively. This includes a detection in the SRM launcher file to see if the config has been updated with this feature (if the user has reset/updated configs in EmuDeck). If not, falls back to old zenity prompt. * Cleaned up Steam ROM Manager script

+++ tools/launchers/srm/steamrommanager.sh
@@ -12,7 +12,16 @@ else
 	SRM_init
 fi 
 
-zenity --question \
-	--width 450 \
-	--title "Close Steam/Steam Input?" \
-	--text "Exit Steam to launch Steam ROM Manager? Desktop controls will revert to Lizard Mode until Steam is reopened. Use L2/R2 to click and the trackpad to move the cursor." && (kill -15 $(pidof steam) & "$SRM_toolPath")
+if grep -q '"autoKillSteam": true' "$HOME/.config/steam-rom-manager/userData/userSettings.json"; then
+	echo "Steam ROM Manager path: $SRM_toolPath"
+	echo "autoKillSteam set to true in Steam ROM Manager. Skipping zenity prompt."
+	"$SRM_toolPath"
+else
+	echo "Steam ROM Manager path: $SRM_toolPath"
+	echo "autoKillSteam set to false in Steam ROM Manager. Loading zenity prompt."
+	zenity --question \
+		--width 450 \
+		--title "Close Steam/Steam Input?" \
+		--text "Exit Steam to launch Steam ROM Manager? Desktop controls will revert to Lizard Mode until Steam is reopened. Use L2/R2 to click and the trackpad to move the cursor." && (kill -15 $(pidof steam) & "$SRM_toolPath")
+fi
+
